Fixed issues and changed to use primary zone ID
Changed it to use primary zone ID and fixed issues with scanning dats for mobs and view TOD's.

+++ DataLibrary/Classes/CampedMob.vb
@@ -21,7 +21,7 @@
     
     Public Property ServerID() As Long
     
-    Public Property Zone() As Byte
+    Public Property Zone() As Short
     
     Public Property IsDead() As Boolean
